% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{adj_from_events}
\alias{adj_from_events}
\title{Extract an adjacency matrix from an event list}
\usage{
adj_from_events(event_list, node_labels = NULL, from = 3, check_errors = FALSE)
}
\arguments{
\item{event_list}{The event list, containing at least a sender ID column and
columns containing binary dummy variables for each character.}

\item{node_labels}{An optional character vector containing node names used to
label rows and columns in the adjacency matrix. If not provided, names are
inferred from the recipient dummy column names of the event list.}

\item{from}{The column containing the sender IDs, followed by dummy
variables for each character.}

\item{check_errors}{If TRUE, the function checks the event list for common
data input errors. It looks for characters with self-ties, empty rows where
no recipient was inputted, and any values other than 1 and 0 in the
recipient dummy columns. Returns a message reporting the results of the
checks.}
}
\value{
An adjacency matrix aggregated from interactions in an event list.
}
\description{
A helper function for extracting an adjacency matrix based on
  the interactions in an event list.

  The event list should be structured consistently with the format
  specified elsewhere in this package documentation (at least a sender ID
  column followed by columns containing binary dummy variables for each
  character).
}
\examples{
tfa <- movienetdata::starwars_01
tfa_adj <- adj_from_events(event_list = tfa$event_list,
                           node_labels = tfa$node_list$char_name,
                           check_errors = TRUE)

}
