% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charinet.R
\docType{package}
\name{charinet}
\alias{charinet}
\title{charinet: A Package for Character Interaction Network Analysis}
\description{
This package contains a number of functions for working with character
interaction network data. The project is designed to work with time-ordered
movie dialogue data, but many functions can work on any data consisting of
similarly structured event lists.
}
\details{
For more information visit \url{https://github.com/pj398/charinet}.
}
