% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{check_for_errors}
\alias{check_for_errors}
\title{Check for easily identifiable data input errors}
\usage{
check_for_errors(adjacency = NULL, event_list = NULL, start_at = 4)
}
\arguments{
\item{adjacency}{The adjacency matrix returned from data read-in.}

\item{event_list}{The event list returned from data read-in.}

\item{start_at}{Where to start checking for unexpected values. Important to
note that unlike other functions in the package, start_at should be set to
the first recipient dummy column in the data, which is set by default to 4.}
}
\value{
A message reportinf the results of the check.
}
\description{
This helper function checks for common data input errors. It
  looks for characters with self-ties, empty rows where no recipient was
  inputted, and any values other than 1 and 0.
}
