% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{hm_tidy}
\alias{hm_tidy}
\title{Convert heatmap matrix to tidy format}
\usage{
hm_tidy(input_heatmap)
}
\arguments{
\item{input_heatmap}{The character-by-chunk heatmap matrix produced by the
\code{story_heatmap} function.}
}
\value{
A tibble.
}
\description{
A simple helper function which converts the character-by-chunk
  heatmap matrix into a tidy data format which plays nicely with ggplot2 and
  other tidyverse packages.
}
\examples{
tfa <- movienetdata::starwars_01
my_heatmap <- story_heatmap(event_list = tfa$event_list,
                            char_names = tfa$node_list$char_name,
                            story_chunk_col = 2,
                            from = 3)
my_tidy_heatmap <- hm_tidy(my_heatmap)

}
