% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movienet.R
\docType{package}
\name{movienet}
\alias{movienet}
\title{movienet: A Package for Basic Character Interaction Network Analysis}
\description{
This package contains a number of functions for working with character
interaction network data. The project is designed to work with time-ordered
movie dialogue data, but many functions can work on any data consisting of
similarly structured event lists.
}
\details{
For more information visit \url{https://github.com/pj398/movienet}.
}
