% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrative_centrality.R
\name{nc_tidy}
\alias{nc_tidy}
\title{Convert narrative centrality scores to tidy format}
\usage{
nc_tidy(input_scores)
}
\arguments{
\item{input_scores}{The N-by-T matrix exported from the
\code{narrative_centrality} function.}
}
\value{
A tibble.
}
\description{
A simple helper function which converts the N-by-T matrix of
  narrative centrality scores into a tidy data format which plays nicely with
  ggplot2 and other tidyverse packages.
}
\examples{
tfa <- movienetData::starwars_01
tfa_scores <- narrative_centrality(tfa$event_list,
                                   chars = tfa[[2]]$char_name,
                                   wp = 0.01,
                                   start_at = 3)
my_tidy_scores <- nc_tidy(tfa_scores$out_scores)

}
