% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_charinet.R
\name{plot_charinet}
\alias{plot_charinet}
\title{Plot the character interaction network using ggraph}
\usage{
plot_charinet(
  adjacency = NULL,
  char_names = NULL,
  degree = NULL,
  cutoff = NULL,
  drop_isolates = TRUE,
  node_fill = "#eb4034",
  parallel_edges = FALSE,
  title = ""
)
}
\arguments{
\item{adjacency}{The adjacency matrix based on a given character interaction
network event list.}

\item{char_names}{A vector of character names, arranged in order of the
character IDs. If not supplied, names will be inferred from the rownames of
\code{adjacency}.}

\item{degree}{A vector of numeric values containing information on the
degree/activity of nodes. This will be used to scale the size of the nodes
in the network diagram.}

\item{cutoff}{A numeric value to be used to filter the visualisation to only
include those characters whose \code{degree} is greater than \code{cutoff}.}

\item{drop_isolates}{Logical. If TRUE, nodes with no ties to other nodes will
not be included in the graph.}

\item{node_fill}{A character string providing the colour to be used to fill
the nodes.}

\item{parallel_edges}{Logical. If TRUE, directed edges will be drawn as two
separate parallel arrows. If FALSE, a single double-sided arrow will be
drawn.}

\item{title}{An optional character string which can be used to add a title to
the visualisation.}
}
\value{
A ggplot2 plot.
}
\description{
Create a quick visualisation of the aggregated character
  interaction network.
}
\examples{
tfa <- movienetdata::starwars_01
\dontrun{
# Plot the network, dropping characters who don't speak more than 3 lines
plot_charinet(adjacency = tfa$adjacency,
              char_names = tfa$nodelist$char_name,
              degree = tfa$nodelist$nlines,
              cutoff = 3,
              title = "Dialogue interactions in Star Wars: The Force Awakens")}

}
