% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot heatmap using ggplot2}
\usage{
plot_heatmap(input_heatmap)
}
\arguments{
\item{input_heatmap}{The tidy-formatted heatmap data, as produced by passing
the matrix returned by the \code{story_heatmap} function through the
\code{hm_tidy} function.}
}
\value{
A ggplot2 plot.
}
\description{
Create a quick heatmap visualisation of character activity by
  story chunks using ggplot2.
}
\examples{
tfa <- movienetdata::starwars_01
story_heatmap(tfa$event_list,
              char_names = tfa$node_list$char_name,
              n_chunks = 4) \%>\%
  hm_tidy() \%>\%
  plot_heatmap()

}
