% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrative_centrality.R
\name{plot_nc}
\alias{plot_nc}
\title{Plot narrative centrality scores using ggplot2}
\usage{
plot_nc(input_scores, label_all = FALSE)
}
\arguments{
\item{input_scores}{The N-by-T matrix exported from the
\code{narrative_centrality} function.}

\item{label_all}{Logical. If TRUE, each character in the line graph will be
labelled. If FALSE, only those characters whose normalised scores at the
end of the narrative are higher than their normalised scores at the start
are labelled (if input scores are not normalised, this won't have any
effect).}
}
\value{
A ggplot2 plot.
}
\description{
A simple helper function which converts the N-by-T matrix of
  narrative centrality scores into a tidy data format which plays nicely with
  ggplot2 and other tidyverse packages.
}
\examples{
tfa <- movienetdata::starwars_01
tfa_scores <- narrative_centrality(tfa$event_list,
                                   char_names = tfa[[2]]$char_name,
                                   wp = 0.01,
                                   from = 3)
plot_nc(tfa_scores$out_scores)

}
